Extract Hive blockchain integration into HiveUtil.
Move RPC failover, account validation, and signature checks out of PlayerUtil into a dedicated class with its own unit tests.
