+++ includes/classes/PlayerUtil.php
@@ -8,7 +8,6 @@ use HiveNova\Core\Cache;
 use HiveNova\Core\Universe;
 use HiveNova\Core\FleetFunctions;
 use Exception;
-use Hive\Hive;
 
 /**
  *  2Moons 
@@ -25,49 +24,8 @@ use Hive\Hive;
  * @link https://github.com/jkroepke/2Moons
  */
 
-// import hive-php and instantiate $hive
-$hivePhp = __DIR__.'/../../vendor/mahdiyari/hive-php/lib/Hive.php';
-if (file_exists($hivePhp)) {
-    require_once $hivePhp;
-}
-
 class PlayerUtil
 {
-	static public function getHiveRpcNodes(): array
-	{
-		return HIVE_RPC_NODES;
-	}
-
-	static public function isHiveRpcError(mixed $result): bool
-	{
-		if (!is_array($result)) {
-			return true;
-		}
-
-		return array_key_exists('code', $result) && array_key_exists('message', $result);
-	}
-
-	static public function hiveRpcCall(string $method, string $params): mixed
-	{
-		foreach (PlayerUtil::getHiveRpcNodes() as $rpcNode) {
-			try {
-				$hive = new Hive([
-					'rpcNodes' => [$rpcNode],
-					'timeout'  => HIVE_RPC_TIMEOUT,
-				]);
-				$result = $hive->call($method, $params);
-			} catch (\Throwable $e) {
-				continue;
-			}
-
-			if (!PlayerUtil::isHiveRpcError($result)) {
-				return $result;
-			}
-		}
-
-		return null;
-	}
-
 	static public function cryptPassword($password)
 	{
 		return password_hash((string) $password, PASSWORD_BCRYPT, ['cost' => 13]);
@@ -104,76 +62,9 @@ class PlayerUtil
 		}
 	}
 
-	static public function isHiveAccountValid($hiveaccount)
-	{
-		if (is_null($hiveaccount) || strlen($hiveaccount) == 0 || strlen((string) $hiveaccount) > 16) {
-			return false;
-		}
-		return preg_match('/^[a-z][-a-z0-9]+[a-z0-9](\.[a-z][-a-z0-9]+[a-z0-9])*$/', (string) $hiveaccount);
-	}
-
-	static public function isHiveSignValid($hiveaccount, $signedblob)
-	{
-		// verify account
-		if (!PlayerUtil::isHiveAccountValid($hiveaccount)) {
-			return false;
-		}
-
-		// verify length
-		if (is_null($signedblob) || strlen($signedblob) < 32 || strlen($signedblob) > 132) {
-			return false;
-		}
-		// verify content
-		if (!PlayerUtil::isNameValid($signedblob)) {
-			return false;
-		}
-
-		$result = PlayerUtil::hiveRpcCall('condenser_api.get_accounts', '[["'.$hiveaccount.'"]]');
-
-		if(!is_array($result) || count($result) == 0 || !isset($result[0]) || !array_key_exists('posting',$result[0])) {
-			return false;
-		}
-
-		$publicKeyString = $result[0]['posting']['key_auths'][0][0];
-		$publicKey = (new Hive())->publicKeyFrom($publicKeyString);
-
-		if (is_null($publicKey)) {
-			return false;
-		}
-
-		$message = hash('sha256', $hiveaccount.' is my account.');
-		try {
-			$verified = $publicKey->verify($message, $signedblob);
-		} catch (\Throwable $e) {
-			return false;
-		}
-		if ($verified) {
-			return true;
-		}
-
-		return false;
-	}
-
-	static public function isHiveAccountExists($hiveaccount) {
-		$hiveaccount = strtolower((string) $hiveaccount); // force lower case
-
-		// verify account
-		if (!PlayerUtil::isHiveAccountValid($hiveaccount)) {
-			return false;
-		}
-
-		$result = PlayerUtil::hiveRpcCall('condenser_api.get_accounts', '[["'.$hiveaccount.'"]]');
-
-		if(is_array($result) && count($result) > 0) {
-			return true;
-		}
-
-		return false;
-	}
-
 	static public function getPlayerAvatarURL($USER){
 		$usernameLower = strtolower((string) $USER['username']);
-		if (PlayerUtil::isHiveAccountValid($usernameLower) && isset($USER['hive_account']) && $usernameLower === $USER['hive_account']) {
+		if (HiveUtil::isAccountValid($usernameLower) && isset($USER['hive_account']) && $usernameLower === $USER['hive_account']) {
 			return 'https://images.hive.blog/u/'.$usernameLower.'/avatar';
 		}
 
@@ -182,7 +73,7 @@ class PlayerUtil
 
 	static public function getPlayerBadges($USER){
 		$usernameLower = strtolower((string) $USER['username']);
-		if (PlayerUtil::isHiveAccountValid($usernameLower) && isset($USER['hive_account']) && $usernameLower === $USER['hive_account']) {
+		if (HiveUtil::isAccountValid($usernameLower) && isset($USER['hive_account']) && $usernameLower === $USER['hive_account']) {
 			return '<a href="https://peakd.com/@'.$USER['hive_account'].'" target="_blank">♦️</a>';
 		}
